Added endpoints for getting pending received requests for posts and s…

+++ paws-back/src/main/java/com/example/pawsback/paws/post/PostRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface PostRepository extends JpaRepository<Post, Long> {
     List<Post> findPostByUserId(Long id);
     Post findPostByPetName(String petName);
-    Post findPostById(int postId);
+    Post findPostById(Long postId);
     @Query("SELECT u FROM Post u WHERE u.isAdopted = false")
     List<Post> findAllNonAdoptedPosts();
     @Query("SELECT p FROM Post p WHERE p.isAdopted = false " + "AND (:minAge IS NULL OR p.age >= :minAge) " + "AND (:maxAge IS NULL OR p.age <= :maxAge) " + "AND (:sex IS NULL OR p.sex = :sex) " + "AND (:race IS NULL OR p.race = :race)")
